Feat: logged in user와 연결 및 예외처리

+++ src/main/java/com/sparta/plantdiary/service/CommentService.java
@@ -4,7 +4,9 @@ import com.sparta.plantdiary.command.CreateCommentCommand;
 import com.sparta.plantdiary.command.UpdateCommentCommand;
 import com.sparta.plantdiary.entity.Comment;
 import com.sparta.plantdiary.entity.Post;
+import com.sparta.plantdiary.error.ForbiddenException;
 import com.sparta.plantdiary.error.NotFoundException;
+import com.sparta.plantdiary.jwt.TokenProvider;
 import com.sparta.plantdiary.repository.CommentRepository;
 import com.sparta.plantdiary.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
@@ -18,6 +20,8 @@ public class CommentService {
 
     private final CommentRepository commentRepository;
     private final PostRepository postRepository;
+    private final TokenProvider tokenProvider;
+
 
     public Comment create(CreateCommentCommand command) throws NotFoundException {
         Post post = postRepository.get(command.getPostId()).orElseThrow(() -> new NotFoundException("게시글을 찾을 수 없습니다."));
@@ -31,10 +35,14 @@ public class CommentService {
         return comments;
     }
 
-    public Comment updateById(UpdateCommentCommand command) throws NotFoundException {
+    public Comment updateById(UpdateCommentCommand command) throws NotFoundException, ForbiddenException {
         postRepository.get(command.getPostId()).orElseThrow(() -> new NotFoundException("게시글을 찾을 수 없습니다."));
 
         Comment comment = commentRepository.findById(command.getId()).orElseThrow(() -> new NotFoundException("게시글을 찾을 수 없습니다."));
+
+        if (comment.getWriter().getId() != tokenProvider.getMemberFromAuthentication().getId()) {
+            throw new ForbiddenException("접근할 수 없습니다.");
+        }
 
         comment.setContent(command.getContent());
         commentRepository.save(comment);
@@ -42,8 +50,13 @@ public class CommentService {
         return comment;
     }
 
-    public void deleteById(Long id) throws NotFoundException {
-        commentRepository.findById(id).orElseThrow(() -> new NotFoundException("게시글을 찾을 수 없습니다."));
+    public void deleteById(Long id) throws NotFoundException, ForbiddenException {
+        Comment comment = commentRepository.findById(id).orElseThrow(() -> new NotFoundException("게시글을 찾을 수 없습니다."));
+
+        if (comment.getWriter().getId() != tokenProvider.getMemberFromAuthentication().getId()) {
+            throw new ForbiddenException("접근할 수 없습니다.");
+        }
+
         commentRepository.deleteById(id);
     }
